Move junit callback plugin out of base
The junit callback plugin is not useful for new end users so it doesn't
really fit the mandate of ansible-base.  However, it is needed for
ansible-test.  We need to consider how developers wanting to run
ansible-test will get the junit plugin do determine if we can remove
this or not.