Restore `debug` and `host_pinned` strategies.
Both of these strategy plugins are nothing but documentation. All functionality is implemented by the underlying `linear` and `free` strategies, and in the case of `debug`, by `lib/ansible/strategy/__init__.py` itself.

If we truly want to remove these plugins from ansible-base, the code should first be extracted from the base classes into the strategy plugins.

It's also worth noting that there are multiple integration tests for ansible-base that depend on both of these plugins.